refactor: Change DeleteUser API endpoint implementation and unify return messages

+++ Server/DelTSZ/Controllers/AuthController.cs
@@ -1,7 +1,7 @@
+using System.ComponentModel.DataAnnotations;
 using DelTSZ.Models.Users;
 using DelTSZ.Services.Authentication;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 
 namespace DelTSZ.Controllers;
@@ -11,102 +11,89 @@ namespace DelTSZ.Controllers;
 public class AuthController(IAuthService authService) : ControllerBase
 {
     [HttpPost("register")]
-    public async Task<IActionResult> Register(Registration request)
+    public async Task<IActionResult> Register([Required] Registration registration)
     {
-        IdentityResult result;
-
         try
         {
-            result = await authService.RegisterCostumer(request);
+            var result = await authService.RegisterCostumer(registration);
 
             if (!result.Succeeded)
-            {
-                return BadRequest(result);
-            }
+                return Conflict(result);
+
+            return Ok(new { message = "Register successfully.", result });
         }
         catch (Exception)
         {
-            return BadRequest("Something went wrong, please try again.");
+            return BadRequest(new { message = "Something went wrong, please try again." });
         }
-
-        return Ok(new { message = "Register successfully.", result });
     }
 
     [HttpPost("register/producer"), Authorize(Roles = "Owner")]
-    public async Task<IActionResult> RegisterProducer(Registration request)
+    public async Task<IActionResult> RegisterProducer([Required] Registration registration)
     {
-        IdentityResult result;
-
         try
         {
-            result = await authService.RegisterProducer(request);
+            var result = await authService.RegisterProducer(registration);
 
             if (!result.Succeeded)
-            {
-                return BadRequest(result);
-            }
+                return Conflict(result);
+
+            return Ok(new { message = "Register successfully.", result });
         }
         catch (Exception)
         {
-            return BadRequest("Something went wrong, please try again.");
+            return BadRequest(new { message = "Something went wrong, please try again." });
         }
-
-        return Ok(new { message = "Register successfully.", result });
     }
 
-    [HttpDelete, Authorize(Roles = "Producer, Costumer")]
-    public async Task<IActionResult> DeleteUser()
+    [HttpPost("delete"), Authorize(Roles = "Producer, Costumer")]
+    public async Task<IActionResult> DeleteUser([FromBody] string password)
     {
         try
         {
             var id = HttpContext.User.Claims.FirstOrDefault(c => c.Type.Contains("identifier"))?.Value;
             if (id == null)
-                return Conflict("Wrong user id.");
+                return Conflict(new { message = "Wrong user id." });
 
             var user = await authService.FindUserById(id);
             if (user == null)
-                return NotFound("User not found.");
+                return NotFound(new { message = "User not found." });
+
+            if (!await authService.CheckPassword(user, password))
+                return Conflict(new { message = "Incorrect password." });
 
             var result = await authService.DeleteUser(user);
-
             if (!result.Succeeded)
-            {
-                return BadRequest(result);
-            }
+                return Conflict(result);
 
-            return Ok(new { message = "Delete successful.", result });
+            await authService.Logout();
+            return Ok(new { message = "Delete and logout successful.", result });
         }
         catch (Exception)
         {
-            return BadRequest("Something went wrong, please try again.");
+            return BadRequest(new { message = "Something went wrong, please try again." });
         }
     }
 
     [HttpPost("login")]
-    public async Task<IActionResult> Login(Login login)
+    public async Task<IActionResult> Login([Required] Login login)
     {
         try
         {
             var user = await authService.FindUserByEmail(login.Email);
-
             if (user == null)
-            {
-                return Unauthorized("Check your login credentials and try again");
-            }
+                return Unauthorized(new { message = "Check your login credentials and try again." });
 
             var result = await authService.Login(user, login.Password);
-
             if (!result.Succeeded)
-            {
-                return Unauthorized("Check your login credentials and try again");
-            }
+                return Unauthorized(new { message = "Check your login credentials and try again." });
+
+            return Ok(new { message = "Login successful." });
         }
         catch (Exception)
         {
-            return BadRequest("Something went wrong, please try again.");
+            return BadRequest(new { message = "Something went wrong, please try again." });
         }
-
-        return Ok(new { message = "Login successful." });
     }
 
     [HttpGet("logout"), Authorize]
@@ -115,12 +102,11 @@ public class AuthController(IAuthService authService) : ControllerBase
         try
         {
             await authService.Logout();
+            return Ok(new { message = "Logout successful." });
         }
         catch (Exception)
         {
-            return BadRequest("Something went wrong, please try again.");
+            return BadRequest(new { message = "Something went wrong, please try again." });
         }
-
-        return Ok(new { message = "Logout successful." });
     }
 }
